feat : 모델에 이슈를 저장하기 위한 메서드 생성 및 이슈가 생성될 때마다 모델db에 이슈 등록 유사도 점수는 long타입으로 변경

+++ src/main/java/com/se/its/domain/issue/dto/response/IssueRecommendResponseDto.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class IssueRecommendResponseDto {
     private Long issue_id;
-    private Double score;
+    private Long score;
 }
